move datastore items to separte file similar to status
